Changed the way the cropped is done.
Previously we used a savebox to make the final PDF contain the same
bounding box as the original ESP file. However, this approach cuts any part
of the psfrag replacements that go outside of the original bounding box,
since latex determines the bounding box before the psfrag replacements take
place..

Now we don't reduce the white space around the picture directly in latex
but instead rely on the external program "pdfcrop" to crop the white space
after the figure is generated.

We now rely on the ext